Added recommendations for every movie and global mappers for movie and movie search

+++ app/src/main/java/com/sychev/movieapp/domain/model/Movie.kt
@@ -24,7 +24,7 @@ data class Movie(
     val tagline: String? = null,
     val title: String? = null,
     val voteAverage: Float? = null,
-    val voteCount: Float? = null,
+    val voteCount: Int? = null,
     var watchStatus: Boolean? = null,
 ): Parcelable {
     @Parcelize
